requirements.txt
streamlit>=1.45.0
azure-ai-projects==1.0.0b10
azure-identity>=1.20.0
python-dotenv>=1.1.0
